requirement.txt
fastapi==0.104.1
uvicorn==0.24.0
asyncpg==0.29.0
python-multipart==0.0.6
aiofiles==23.2.1
PyJWT==2.8.0